+++ lib/simmer.rb
@@ -22,7 +22,7 @@ require 'set'
 require 'stringento'
 require 'yaml'
 
-# Monkey-patching core libaries
+# Monkey-patching core libraries
 require_relative 'simmer/core_ext/hash'
 Hash.include Simmer::CoreExt::Hash
 
@@ -30,6 +30,7 @@ Hash.include Simmer::CoreExt::Hash
 require_relative 'simmer/util'
 
 # Core code
+require_relative 'simmer/bootstrap'
 require_relative 'simmer/configuration'
 require_relative 'simmer/database'
 require_relative 'simmer/externals'
@@ -39,11 +40,7 @@ require_relative 'simmer/specification'
 require_relative 'simmer/suite'
 
 # The main entry-point API for the library.
-
-# TODO: split this up and re-enable this cop:
-# rubocop:disable Metrics/ModuleLength
 module Simmer
-  # rubocop:enable Metrics/ModuleLength
   DEFAULT_CONFIG_PATH = File.join('config', 'simmer.yaml').freeze
   DEFAULT_SIMMER_DIR  = 'simmer'
 
@@ -54,44 +51,19 @@ module Simmer
       out: $stdout,
       simmer_dir: DEFAULT_SIMMER_DIR
     )
-      configuration = make_configuration(config_path: config_path, simmer_dir: simmer_dir)
-      specs         = make_specifications(path, configuration.tests_dir)
-      out_router    = make_output_router(configuration, out)
-      runner        = make_runner(configuration, out_router)
-      suite         = make_suite(configuration, out, runner)
-
-      suite.run(specs)
+      Bootstrap.new(
+        spec_path: path,
+        config_path: config_path,
+        simmer_dir: simmer_dir,
+        callback_configuration: callback_configuration,
+        console_out: out
+      ).run_suite
     end
 
-    def make_configuration(config_path: DEFAULT_CONFIG_PATH, simmer_dir: DEFAULT_SIMMER_DIR)
-      raw_config = yaml_reader.smash(config_path)
-      Configuration.new(raw_config, simmer_dir, callbacks: callback_configuration)
+    def configuration(config_path: DEFAULT_CONFIG_PATH, simmer_dir: DEFAULT_SIMMER_DIR)
+      Bootstrap.new(config_path: config_path, simmer_dir: simmer_dir).configuration
     end
-
-    def callback_configuration
-      @callback_configuration ||= Configuration::CallbackDsl.new
-    end
-
-    def make_runner(configuration, out_router)
-      database     = make_mysql_database(configuration)
-      file_system  = make_file_system(configuration)
-      fixture_set  = make_fixture_set(configuration)
-      spoon_client = make_spoon_client(configuration)
-
-      runner = Runner.new(
-        database: database,
-        file_system: file_system,
-        fixture_set: fixture_set,
-        out: out_router,
-        spoon_client: spoon_client
-      )
-
-      ReRunner.new(
-        runner,
-        out_router,
-        timeout_failure_retry_count: configuration.timeout_failure_retry_count
-      )
-    end
+    alias make_configuration configuration
 
     def configure(&block)
       # TODO: support the arity 1 case
@@ -107,83 +79,8 @@ module Simmer
 
     private
 
-    def yaml_reader
-      Util::YamlReader.new
-    end
-
-    def make_specifications(path, tests_dir)
-      path = path.to_s.empty? ? tests_dir : path
-
-      Util::YamlReader.new.read(path).map do |file|
-        config = (file.data || {}).merge(path: file.path)
-
-        Specification.make(config)
-      end
-    end
-
-    def make_fixture_set(configuration)
-      config = Util::YamlReader.new.smash(configuration.fixtures_dir)
-
-      Database::FixtureSet.new(config)
-    end
-
-    def make_mysql_database(configuration)
-      config         = configuration.mysql_database_config.symbolize_keys
-      client         = Mysql2::Client.new(config)
-      exclude_tables = config[:exclude_tables]
-
-      Externals::MysqlDatabase.new(client, exclude_tables)
-    end
-
-    def make_file_system(configuration)
-      if configuration.aws_file_system?
-        make_aws_file_system(configuration)
-      elsif configuration.local_file_system?
-        make_local_file_system(configuration)
-      else
-        raise ArgumentError, 'cannot determine file system'
-      end
-    end
-
-    def make_aws_file_system(configuration)
-      config      = configuration.aws_file_system_config.symbolize_keys
-      client_args = config.slice(:access_key_id, :secret_access_key, :region)
-      client      = Aws::S3::Client.new(client_args)
-
-      Externals::AwsFileSystem.new(
-        client,
-        config[:bucket],
-        config[:encryption],
-        configuration.files_dir
-      )
-    end
-
-    def make_local_file_system(configuration)
-      config = configuration.local_file_system_config.symbolize_keys
-
-      Externals::LocalFileSystem.new(config[:dir], configuration.files_dir)
-    end
-
-    def make_spoon_client(configuration)
-      config     = (configuration.spoon_client_config || {}).symbolize_keys
-      spoon_args = config.slice(:args, :dir, :kitchen, :pan, :timeout_in_seconds)
-      spoon      = Pdi::Spoon.new(spoon_args)
-
-      Externals::SpoonClient.new(configuration.files_dir, spoon)
-    end
-
-    def make_suite(configuration, out, runner)
-      Suite.new(
-        config: configuration,
-        out: out,
-        results_dir: configuration.results_dir,
-        runner: runner
-      )
-    end
-
-    def make_output_router(configuration, console_out)
-      pdi_out = Suite::PdiOutputWriter.new(configuration.results_dir)
-      Simmer::Suite::OutputRouter.new(console_out, pdi_out)
+    def callback_configuration
+      @callback_configuration ||= Configuration::CallbackDsl.new
     end
   end
 end